// src/interfaces/IProducer.ts
/**
 * Interface for producer
 */
export interface IProducer {
    country: string
    district: string
    district_jpn: string
    family_head: string
    family_head_jpn: string
    fermentation_detail: string
    field_area: string
    field_detail: string
    founded_year: string
    headquarter: string
    headquarter_jpn: string
    history_detail: string
    home_page: string
    id: string
    importer: string
    is_multi: string
    is_original: string
    name: string
    name_jpn: string
    original_contents: string
    region: string
    region_jpn: string
    short_name: string
    short_name_jpn: string
    village: string
    village_jpn: string
}